LPS-76650 Move to internal
