+++ modules/apps/twitter/twitter-service/src/main/java/com/liferay/twitter/internal/util/TimelineProcessorUtil.java
@@ -12,16 +12,31 @@
  * details.
  */
 
-package com.liferay.twitter.util;
+package com.liferay.twitter.internal.util;
 
 import com.liferay.portal.kernel.json.JSONArray;
 
 /**
  * @author Shinn Lok
+ * @author Peter Fellwock
  */
-public interface TimelineProcessor {
+public class TimelineProcessorUtil {
 
-	public JSONArray getUserTimelineJSONArray(
-		String twitterScreenName, long sinceId);
+	public static TimelineProcessor getInstance() {
+		if (_timelineProcessor == null) {
+			_timelineProcessor = new HttpTimelineProcessor();
+		}
+
+		return _timelineProcessor;
+	}
+
+	public static JSONArray getUserTimelineJSONArray(
+		String twitterScreenName, long sinceId) {
+
+		return getInstance().getUserTimelineJSONArray(
+			twitterScreenName, sinceId);
+	}
+
+	private static TimelineProcessor _timelineProcessor;
 
 }
